chore(libft): Use structs instead of pointers to struct in bstr_cmp
This will make a few things simpler because we can always dereference one of the
arguments if it is a pointer, on the contrary, you can't take the address of an
rvalue (eg. str_to_bstr()).

Now we can do:
```
t_bstr *str;
str = something();
bstr_cmp(str_to_bstr("hello"), *str);
```